depth is now integer, fix when passing depth as parameter in the recursion
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/triva@465 96ca98c3-5b99-4e05-a768-75e2d8bf8841

+++ trunk/src/time-slice/Treemap.h
@@ -5,23 +5,23 @@
 {
 	float width, height;
 	float x, y;
-	float depth;
+	int depth;
 }
 - (float) width;
 - (float) height;
 - (float) x;
 - (float) y;
-- (float) depth;
+- (int) depth;
 
 - (void) setWidth: (float) w;
 - (void) setHeight: (float) h;
 - (void) setX: (float) xp;
 - (void) setY: (float) yp;
-- (void) setDepth: (float) d;
+- (void) setDepth: (int) d;
 
 - (void) calculateWithWidth: (float) w andHeight: (float) h;
 - (void) calculateWithWidth: (float) W
               height: (float) H
               factor: (float) factor
-                depth: (float) d;
+                depth: (int) d;
 @end
